+++ config.cpp
@@ -1,4 +1,10 @@
-<?xml version="1.0" encoding="utf-8"?>
-<project name="SagaTheHorde">
-    <define name="COMPONENT_NAME" value="SagaTheHorde"/>
-</project>
+class CfgPatches
+{
+	class SagaTheHorde
+	{
+		units[] = {};
+		weapons[] = {};
+		requiredVersion = 0.1;
+		requiredAddons[] = {"DZ_Data"};
+	};
+};